Added manual option for location qualification and added user change
